# Conflicts:
#	base/threading/platform_thread_mac.mm
#	chrome/browser/ui/views/frame/browser_non_client_frame_view_mac.mm
#	components/remote_cocoa/app_shim/browser_native_widget_window_mac.mm
#	content/app_shim_remote_cocoa/web_contents_view_cocoa.mm
#	content/browser/web_contents/web_drag_dest_mac.mm
#	device/fido/mac/authenticator.mm
#	net/base/platform_mime_util_mac.mm

+++ chrome/browser/ash/login/users/chrome_user_manager_impl.cc
@@ -52,6 +52,7 @@
 #include "chrome/browser/ash/login/users/multi_profile_user_controller.h"
 #include "chrome/browser/ash/login/users/supervised_user_manager_impl.h"
 #include "chrome/browser/ash/policy/core/browser_policy_connector_ash.h"
+#include "chrome/browser/ash/policy/core/device_local_account.h"
 #include "chrome/browser/ash/policy/external_data/handlers/crostini_ansible_playbook_external_data_handler.h"
 #include "chrome/browser/ash/policy/external_data/handlers/preconfigured_desk_templates_external_data_handler.h"
 #include "chrome/browser/ash/policy/external_data/handlers/print_servers_external_data_handler.h"
@@ -252,6 +253,38 @@ void CheckProfileForSanity() {
   CHECK(
       !g_browser_process->profile_manager()->GetProfileByPath(user_profile_dir))
       << "The user profile was loaded before we mounted the cryptohome.";
+}
+
+user_manager::UserManager::EphemeralModeConfig CreateEphemeralModeConfig(
+    ash::CrosSettings* cros_settings) {
+  DCHECK(cros_settings);
+
+  bool ephemeral_users_enabled = false;
+  cros_settings->GetBoolean(ash::kAccountsPrefEphemeralUsersEnabled,
+                            &ephemeral_users_enabled);
+
+  std::vector<AccountId> ephemeral_accounts, non_ephemeral_accounts;
+
+  const auto accounts = policy::GetDeviceLocalAccounts(cros_settings);
+  for (const auto& account : accounts) {
+    switch (account.ephemeral_mode) {
+      case policy::DeviceLocalAccount::EphemeralMode::kEnable:
+        ephemeral_accounts.push_back(AccountId::FromUserEmail(account.user_id));
+        break;
+      case policy::DeviceLocalAccount::EphemeralMode::kDisable:
+        non_ephemeral_accounts.push_back(
+            AccountId::FromUserEmail(account.user_id));
+        break;
+      case policy::DeviceLocalAccount::EphemeralMode::kUnset:
+      case policy::DeviceLocalAccount::EphemeralMode::kFollowDeviceWidePolicy:
+        // Do nothing.
+        break;
+    }
+  }
+
+  return user_manager::UserManager::EphemeralModeConfig(
+      ephemeral_users_enabled, std::move(ephemeral_accounts),
+      std::move(non_ephemeral_accounts));
 }
 
 }  // namespace
@@ -665,10 +698,11 @@ bool ChromeUserManagerImpl::IsUserNonCryptohomeDataEphemeral(
          ChromeUserManager::IsUserNonCryptohomeDataEphemeral(account_id);
 }
 
-bool ChromeUserManagerImpl::AreEphemeralUsersEnabled() const {
+bool ChromeUserManagerImpl::IsEphemeralAccountId(
+    const AccountId& account_id) const {
   policy::BrowserPolicyConnectorAsh* connector =
       g_browser_process->platform_part()->browser_policy_connector_ash();
-  return GetEphemeralUsersEnabled() &&
+  return GetEphemeralModeConfig().IsAccountIdIncluded(account_id) &&
          (connector->IsDeviceEnterpriseManaged() ||
           GetOwnerAccountId().is_valid());
 }
@@ -735,7 +769,7 @@ void ChromeUserManagerImpl::RetrieveTrustedDevicePolicies() {
   if (!GetLocalState())
     return;
 
-  SetEphemeralUsersEnabled(false);
+  SetEphemeralModeConfig(EphemeralModeConfig());
   SetOwnerId(EmptyAccountId());
 
   // Schedule a callback if device policy has not yet been verified.
@@ -746,10 +780,7 @@ void ChromeUserManagerImpl::RetrieveTrustedDevicePolicies() {
     return;
   }
 
-  bool ephemeral_users_enabled = false;
-  cros_settings_->GetBoolean(kAccountsPrefEphemeralUsersEnabled,
-                             &ephemeral_users_enabled);
-  SetEphemeralUsersEnabled(ephemeral_users_enabled);
+  SetEphemeralModeConfig(CreateEphemeralModeConfig(cros_settings_));
 
   std::string owner_email;
   cros_settings_->GetString(kDeviceOwner, &owner_email);
@@ -765,14 +796,15 @@ void ChromeUserManagerImpl::RetrieveTrustedDevicePolicies() {
 
   // If ephemeral users are enabled and we are on the login screen, take this
   // opportunity to clean up by removing all regular users except the owner.
-  if (GetEphemeralUsersEnabled() && !IsUserLoggedIn()) {
+  if (!IsUserLoggedIn()) {
     ScopedListPrefUpdate prefs_users_update(GetLocalState(),
                                             user_manager::kRegularUsersPref);
     prefs_users_update->clear();
     for (user_manager::UserList::iterator it = users_.begin();
          it != users_.end();) {
       const AccountId account_id = (*it)->GetAccountId();
-      if ((*it)->HasGaiaAccount() && account_id != GetOwnerAccountId()) {
+      if ((*it)->HasGaiaAccount() && account_id != GetOwnerAccountId() &&
+          IsEphemeralAccountId(account_id)) {
         user_manager::UserManager::Get()->NotifyUserToBeRemoved(account_id);
         RemoveNonCryptohomeData(account_id);
         DeleteUser(*it);
